features Abono and DetalleVenta

+++ db/schemas/abonoSchema.py
@@ -1,13 +1,13 @@
 from typing import Optional
 from datetime import datetime
 from sqlmodel import SQLModel
-
+from decimal import Decimal
 
 # --- SCHEMAS PARA ABONOS ---
 
 class AbonoBase(SQLModel):
     cliente_id: int
-    monto_abonado: float
+    monto_abonado: Decimal
     notas: Optional[str] = None
 
 class AbonoCreate(AbonoBase):
